LeetCode Q.17. Letter Combinations of a Phone Number
This Question uses back-tracking and is as follows:
Given a string containing digits from 2-9 inclusive, return all possible letter combinations that the number could represent. Return the answer in any order.

A mapping of digits to letters (just like on the telephone buttons) is given below. Note that 1 does not map to any letters.